fix: put lambda functions in utils module to avoid circular imputs
