+++ src/python_library/joint_calc/utils.py
@@ -1,0 +1,10 @@
+import Rhino
+
+
+def volume_lambda(brep):
+    volume = abs(Rhino.Geometry.VolumeMassProperties.Compute(brep).Volume)
+    return volume
+
+
+def surface_lambda(brep):
+    return Rhino.Geometry.AreaMassProperties.Compute(brep).Area
